Test cycle wall-clock cap in StatusProvider
A worker hanging past refresh_timeout_seconds must not stall the
cycle; wait(timeout) returns the done set and stragglers stay absent
from the map (read as UNKNOWN by the loop's elif chain).

+++ src/dragen_align_pa/jobs/ica_status_provider.py
@@ -42,7 +42,7 @@ class ParallelPerIdStatusProvider:
     def __init__(
         self,
         concurrency: int,
-        refresh_timeout_seconds: int,
+        refresh_timeout_seconds: float,
         project_id: str | None = None,
     ) -> None:
         self._concurrency = concurrency
